Add files via upload
Intuition
The code aims to find the middle node of a linked list. It seems to be utilizing a two-pointer approach where one pointer (mid) moves one step at a time while the other pointer (end) moves two steps at a time. When the end pointer reaches the end of the linked list, the mid pointer will be at the middle node.

Approach
The approach involves traversing the linked list with two pointers (mid and end). The mid pointer moves one step at a time, while the end pointer moves two steps at a time. By the time the end pointer reaches the end of the linked list, the mid pointer will be at the middle node.

Complexity
Time complexity:
O(n)
where n is the number of nodes in the linked list. The algorithm traverses the entire linked list once.

Space complexity:
O(1)
since the algorithm only uses a constant amount of extra space for two pointers regardless of the size of the input linked list.

Code
# Definition for singly-linked list.
# class ListNode:
#     def __init__(self, val=0, next=None):
#         self.val = val
#         self.next = next
class Solution:
    def middleNode(self, head: Optional[ListNode]) -> Optional[ListNode]:
        mid = end = head
        while end:
            if end and end.next:
                mid = mid.next
                end = end.next
            end = end.next
        return mid