Add files via upload
Intuition
The problem aims to find the sum of the frequencies of the most frequent elements in a given list of integers.

Approach
The approach employs a systematic process to compute the sum of the frequencies of the most frequent elements within the given list of integers.

Calculate Frequency:
Traverse through the list of integers and maintain a frequency array freq_arr with a size of 101, considering the constraint that the numbers lie within the range of [1, 100]. Increment the frequency count of each encountered element in the array.

Find Maximum Frequency:
During the traversal, keep track of the maximum frequency encountered, denoted by maxf.

Sum of Maximum Frequencies:
For each element, compare its frequency with maxf. If the frequency equals maxf, accumulate its frequency to the sum sumf.

Array Optimization:
Utilize an array (freq_arr) rather than a dictionary or hashmaps to store frequencies. This choice is favorable due to the limited range of integers provided in the problem.

By systematically iterating through the list and maintaining frequency counts, the algorithm efficiently identifies the most frequent elements and computes their cumulative frequencies.

This structured approach ensures optimal performance while handling the given task.

Complexity
Time complexity:
O(n)
where n is the length of the input list nums. The algorithm iterates through the list once to calculate frequencies.

Space complexity:
O(1)
As the size of the frequency array (freq_arr) is fixed regardless of the input size. The additional space used is constant.