Python Code Solution, better approach
Intuition
The problem aims to find the intersection of two lists, which are represented by arrays in this case.

Approach
The approach taken is a straightforward iteration through both arrays. The code compares elements from one array with elements from the other array to identify common elements. If a common element is found, and it's not already in the result list, it's appended to the result list. The code first checks which array is shorter and iterates through the shorter array to optimize the process.

Complexity
Time complexity:
O(m⋅n) or O(n⋅m)
If len(nums1)<len(nums2)len(nums1)<len(nums2), the time complexity is O(m⋅n), where m is the length of nums1 and n is the length of nums2.

If len(nums1)≥len(nums2), the time complexity is O(n⋅m).

Space complexity:
O(k)
where k is the size of the result list.