Python Easy Approach
Intuition
The problem involves counting the number of palindromic substrings in a given string. One intuitive approach could be to expand around each character and check if it forms a palindrome.

Approach
We can define a helper function grow that expands around a center character(s) and counts the palindromic substrings.
Within the main function countSubstrings, we iterate through each character in the string and consider both cases: expanding around a single character and expanding around two adjacent characters.
We accumulate the count of palindromic substrings for each expansion.
Finally, we return the total count of palindromic substrings.
Complexity
Time complexity:
O(n^2)
The grow function takes linear time to expand around each center character, so its time complexity is O(n).
In the countSubstrings function, we iterate through each character in the string, so the overall time complexity is O(n^2).

Space complexity:
O(1)
The space complexity is O(1) as we're not using any additional space proportional to the input size.