Python Code with less space and time complexity
Intuition
The intuition behind this approach is to compute the product of all elements to the left and right of each element in the array, excluding the current element itself. This way, we can calculate the product of all elements except self.

Approach
We initialize a result array res with all elements set to 1. Then, we compute the product of all elements to the left of each element and store it in the res array, updating the prefix value for each iteration. After that, we compute the product of all elements to the right of each element and update the res array accordingly, updating the postfix value for each iteration.

Complexity
Time complexity:
O(n)
where n is the length of the input array nums. We traverse the array twice, once from left to right and once from right to left, each traversal taking linear time.

Space complexity:
O(n)
where n is the length of the input array nums. We use an additional array res of size n to store the final result.